+++ db/migrate/20190215114245_add_offer_type_to_offer.rb
@@ -1,0 +1,5 @@
+class AddOfferTypeToOffer < ActiveRecord::Migration[5.2]
+  def change
+    add_column :offers, :offer_type, :string, null: true
+  end
+end