+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Please view this file on the master branch, on stable branches it's out of date.
   selected (@mkasztelnik)
 - Remove shell command invocation in one of the test (@mkasztelnik)
 - ProjectItem details view crash for open access services (@michal-szostak)
+- Order mailing (@michal-szostak)
 
 ### Security