+++ CHANGELOG.md
@@ -17,9 +17,13 @@ Please view this file on the master branch, on stable branches it's out of date.
 
 ### Fixed
 
-- Small view error in my order detail page 
-
 ### Security
+
+## [1.4.1] - 2019-02-28
+
+### Fixed
+
+- Small view error in my order detail page 
 
 ## [1.4.0] - 2019-02-26